% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbisr.r
\name{orbis.data.filter}
\alias{orbis.data.filter}
\title{Filter tables of Orbis bulk data}
\usage{
orbis.data.filter(orbis.data.path, ...)
}
\arguments{
\item{orbis.data.path}{A path to directory with .rds files containing Orbis specific table from Orbis Bulk Data.}

\item{...}{A filtering conditions to fetch certain rows. (See dplyr::filter)}
}
\value{
A data.table with a subset of a table from Orbis Bulk Data.
}
\description{
Similar to dplyr::filter but for tables of Orbis bulk data saved in multiple .rds files
}
\examples{
none yet...
}
