% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbisr.r
\name{orbis.save.rds}
\alias{orbis.save.rds}
\title{Read orbis raw data table to many .rds files}
\usage{
orbis.save.rds(txt.file.name, txt.file.dir = getwd(),
  data.codes.file = character(0), data.codes.dir = file.path(getwd(),
  "orbis-var-names"), select.codes = character(0), txt.skip.lines = 2,
  txt.nlines = NA, batch.nlines = 10^7,
  batch.file.dir = character(0), batch.file.name = character(0),
  batch.file.name.prefix = "", batch.file.name.sufix = "",
  save.rds = TRUE, return.invisible = FALSE,
  harmonize.cols = character(0), harmonize.progress.by = 10^5,
  harmonize.quite = FALSE, harmonize.procedures = list(list("toascii",
  FALSE), "remove.brackets", "toupper", "apply.nber", "remove.spaces"))
}
\arguments{
\item{txt.file.name}{File name of raw Orbis data}

\item{txt.file.dir}{A path to directory with raw data. Default is working directory.}

\item{data.codes.file}{Name of the .csv file with codes. Default is the same as txt.file.name but with .csv}

\item{data.codes.dir}{A path to .csv file with codes. Default is ./orbis-var-names}

\item{select.codes}{A character vector with felds (code) to select. Default is all.}

\item{txt.skip.lines}{The header of raw data. The default is 2.}

\item{txt.nlines}{Number of lined in the raw data file. Default is calculate with grep}

\item{batch.nlines}{Number of lines to read in batch. The default is 10^7}

\item{batch.file.dir}{Path for saving .rds files. The default is dir same as 'txt.file.name'}

\item{harmonize.cols}{Which columns to harmonize. (Requires harmonizer package.)}

\item{harmonize.progress.by}{(Requires harmonizer package.) Numeric value that is used to split the org.names vector for showing percentage of completion. Default is 0 meaning not to split the vector and thus does not show progress percentage. Designed to be used for long strings.}

\item{harmonize.quite}{(Requires harmonizer package.) Logical value indicating whether or not print messages about procedures progress.}

\item{harmonize.procedures}{(Requires harmonizer package.) List of harmonization procedures. Each procedure can be specified as a string representing procedure name (see details for procedure names) or as a list where the first element should be procedure name (string) and other elements will passed as arguments to this procedure.}
}
\value{
A vector of .rds file names.
}
\description{
Designed to be run from the direcrory where the data is.
}
