% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbisr.r
\docType{package}
\name{orbisr-package}
\alias{orbisr}
\alias{orbisr-package}
\title{Tools for working with Orbis bulk data}
\description{
================================================================================
}
\details{
The only functions you're likely to need from \pkg{orbisr} for deployment of Orbis
database are \code{\link{orbis.save.rds}} and \code{\link{orbis.unrar.txt}}.

For working with deploied Orbis database you will need only \code{\link{orbis.filter}}
}
\author{
\strong{Maintainer}: Stanislav Vlasov \email{s.vlasov@uvt.nl}

}
