% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbisr.r
\name{orbis.save.rds}
\alias{orbis.save.rds}
\title{Read orbis raw data table to many .rds files}
\usage{
orbis.save.rds(orbis.data.raw.file, orbis.data.path = getwd(),
  orbis.data.codes.file = character(0),
  orbis.data.codes.dir = file.path(orbis.data.path, "orbis-var-names"),
  orbis.select.fields = character(0), orbis.data.raw.nlines = 0,
  orbis.data.raw.skip.lines = 2, orbis.batch.nlines = 10^7,
  orbis.batch.path = character(0))
}
\arguments{
\item{orbis.data.raw.file}{File name of raw Orbis data}

\item{orbis.data.path}{A path to directory with raw data. Default is working directory.}

\item{orbis.data.codes.file}{Name of the .csv file with codes. Default is the same as orbis.data.raw.file but with .csv}

\item{orbis.data.codes.dir}{A path to .csv file with codes. Default is ./orbis-var-names}

\item{orbis.select.fields}{A character vector with felds (code) to select. Default is all.}

\item{orbis.data.raw.nlines}{Number of lined in the raw data file. Default is calculate with grep}

\item{orbis.data.raw.skip.lines}{The header of raw data. The default is 2.}

\item{orbis.batch.nlines}{Number of lines to read in batch. The default is 10^7}

\item{orbis.batch.path}{Path for saving .rds files. The default is dir same as 'orbis.data.raw.file'}
}
\value{
A vector of .rds file names
}
\description{
Designed to be run from the direcrory where the data is.
}
\examples{
none yet...
}
